requirements.txt
# Flask==3.0.3
Django==5.0.6
numpy==1.26.4
matplotlib==3.8.4
channels==3.0.5
# essentia==2.1b6.dev1110
essentia-tensorflow==2.1b6.dev1110